handle delete without break the cache

+++ resources/views/components/⚡article-list.blade.php
@@ -14,6 +14,11 @@ new #[\Livewire\Attributes\Title('Mange Articles'),\Livewire\Attributes\Layout('
 
     public function delete(\App\Models\Article $article)
     {
+        if($this->articles->count()<10)
+        {
+            throw new Exception("Nope");
+        }
+
         $article->delete();
         $this->resetPage();
